Reorganized directory to ensure correct imports and more logical structure
