+++ soupy/risk/superquantileRiskMeasureSAA.py
@@ -20,13 +20,12 @@ import scipy.optimize
 from hippylib import ParameterList, Random
 
 from .riskMeasure import RiskMeasure
-from .variables import STATE, PARAMETER, ADJOINT, CONTROL
-from .controlModelHessian import ControlModelHessian
+
+from ..modeling import STATE, PARAMETER, ADJOINT, CONTROL, ControlModelHessian, AugmentedVector
 
 from ..collectives import NullCollective, MultipleSamePartitioningPDEsCollective, \
         MultipleSerialPDEsCollective, allocate_process_sample_sizes
 from .smoothPlusApproximation import SmoothPlusApproximationQuartic, SmoothPlusApproximationSoftplus
-from .augmentedVector import AugmentedVector
